refactor(validation-gen): move list-related validators to list.go
